refactor(react): update radio button tests (#11790)
* refactor(react): update radio button group tests

* refactor(react): update radio button tests

* docs(react): update radio button docs

* test(react): update snapshots
