+++ packages/react/src/components/RadioButton/index.js
@@ -5,13 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import * as FeatureFlags from '@carbon/feature-flags';
-import { default as RadioButtonNext } from './next/RadioButton';
-import { default as RadioButtonClassic } from './RadioButton';
-
-const RadioButton = FeatureFlags.enabled('enable-v11-release')
-  ? RadioButtonNext
-  : RadioButtonClassic;
+import RadioButton from './RadioButton';
 
 export default RadioButton;
-export { default as RadioButtonSkeleton } from './RadioButton.Skeleton';
